extracted polyfill for performance benefits
